Moved navbar into its own component
