+++ src/pages/index.js
@@ -1,5 +1,5 @@
-"use client";
 import React, { useEffect, useState } from "react";
+import Navbar from "./components/Navbar";
 import LandingPage from "./components/LandingPage";
 import About from "./components/About";
 import Guide from "./components/Guide";
@@ -8,166 +8,9 @@ import FAQ from "./components/FAQ";
 import Team from "./components/Team";
 import Background from "./components/Background";
 
-const Navbar = () => {
-  const [isClick, setIsClick] = useState(false);
-
-  const toggleNavbar = () => {
-    setIsClick(!isClick);
-  };
-
-  return (
-    <>
-      <nav className="bg-gray-700 p-5 w-full fixed z-50">
-        <div className="mx-auto px-4 sm:px-6">
-          <div className="flex items-center justify-between h-16">
-            <div className="flex items-center">
-              <div className="flex-shrink-0">
-                <a href="#" className="text-white text-5xl">
-                  sase
-                </a>
-              </div>
-            </div>
-            <div className="flex hidden md:block">
-              <div className="ml-4 flex items-center space-x-6">
-                <a
-                  href="#about"
-                  className="text-white hover:bg-white hover:text-black rounded-lg p-2"
-                >
-                  About
-                </a>
-                <a
-                  href="#key-dates"
-                  className="text-white hover:bg-white hover:text-black rounded-lg p-2"
-                >
-                  Key Dates
-                </a>
-                <a
-                  href="#sponsors"
-                  className="text-white hover:bg-white hover:text-black rounded-lg p-2"
-                >
-                  Sponsors
-                </a>
-                <a
-                  href="#faq"
-                  className="text-white hover:bg-white hover:text-black rounded-lg p-2"
-                >
-                  FAQ
-                </a>
-                <a
-                  href="#scrc-team"
-                  className="text-white hover:bg-white hover:text-black rounded-lg p-2"
-                >
-                  SCRC Team
-                </a>
-              </div>
-            </div>
-            <div className="md:hidden flex items-center">
-              <button
-                className="inline-flex items-center justify center p-2 rounded-md text-white hover:text-white focus:outline-none focus:ring-2 focus:ring-inset focus:ring-white"
-                onClick={toggleNavbar}
-              >
-                {isClick ? (
-                  <svg
-                    className="h-8 w-8"
-                    xmlns="http://www.w3.org/2000/svg"
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    stroke="currentColor"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={2}
-                      d="M6 18L18 6M6 6l12 12"
-                    />
-                  </svg>
-                ) : (
-                  <svg
-                    className="h-8 w-8"
-                    xmlns="http://www.w3.org/2000/svg"
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    stroke="currentColor"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={2}
-                      d="M4 6h16M4 12h16m-7 6h7"
-                    />
-                  </svg>
-                )}
-              </button>
-            </div>
-          </div>
-        </div>
-        {isClick && (
-          <div className="md:hidden">
-            <div className="px-2 pt-2 pb-3 space-y-1 sm:px-3">
-              <a
-                href="#about"
-                className="text-white block hover:bg-white hover:text-black rounded-lg p-2"
-              >
-                About
-              </a>
-              <a
-                href="#key-dates"
-                className="text-white block hover:bg-white hover:text-black rounded-lg p-2"
-              >
-                Key Dates
-              </a>
-              <a
-                href="#sponsors"
-                className="text-white block hover:bg-white hover:text-black rounded-lg p-2"
-              >
-                Sponsors
-              </a>
-              <a
-                href="#faq"
-                className="text-white block hover:bg-white hover:text-black rounded-lg p-2"
-              >
-                FAQ
-              </a>
-              <a
-                href="#scrc-team"
-                className="text-white block hover:bg-white hover:text-black rounded-lg p-2"
-              >
-                SCRC Team
-              </a>
-            </div>
-          </div>
-        )}
-      </nav>
-    </>
-  );
-};
-
 export default function Home() {
-  const [isDesktop, setDesktop] = useState(true);
-
-  const updateMedia = () => {
-    setDesktop(window.innerWidth > 640);
-  };
-
-  useEffect(() => {
-    if (!(typeof window === "undefined")) {
-      window.addEventListener("resize", updateMedia);
-      return () => window.removeEventListener("resize", updateMedia);
-    }
-  }, []);
-
-  const handleCloseModals = () => {
-    var scheduleContainer = document.getElementById("schedule-container");
-    if (scheduleContainer.classList.contains("open-schedule")) {
-      scheduleContainer.classList.remove("open-schedule");
-      scheduleContainer.classList.add("close-schedule");
-      document.body.style.overflow = "auto";
-    }
-  };
-
   return (
     <main>
-      <Navbar />
       <div className="scroll-container">
         <LandingPage />
         <About />
@@ -176,82 +19,7 @@ export default function Home() {
         <FAQ />
         <Team />
       </div>
-      <header className="nav-container">
-        <a className="name-header" href="#">
-          SCRC
-        </a>
-        {isDesktop ? (
-          <nav className="navbar">
-            <a
-              id="about"
-              className="navbar-link"
-              href="#about"
-              onClick={handleCloseModals}
-            >
-              About
-            </a>
-            <a
-              id="key-dates"
-              className="navbar-link"
-              href="#key-dates"
-              onClick={handleCloseModals}
-            >
-              Key Dates
-            </a>
-            <a
-              id="sponsors"
-              className="navbar-link"
-              href="#sponsors"
-              onClick={handleCloseModals}
-            >
-              Sponsors
-            </a>
-            <a
-              id="faq"
-              className="navbar-link"
-              href="#faq"
-              onClick={handleCloseModals}
-            >
-              FAQ
-            </a>
-            <a
-              id="scrc-team"
-              className="navbar-link"
-              href="#scrc-team"
-              onClick={handleCloseModals}
-            >
-              SCRC Team
-            </a>
-          </nav>
-        ) : (
-          <nav className="navbar">
-            <a
-              id="key-dates"
-              className="navbar-link"
-              href="#key-dates"
-              onClick={handleCloseModals}
-            >
-              Key Dates
-            </a>
-            <a
-              id="sponsors"
-              className="navbar-link"
-              href="#sponsors"
-              onClick={handleCloseModals}
-            >
-              Sponsors
-            </a>
-            <a
-              id="faq"
-              className="navbar-link"
-              href="#faq"
-              onClick={handleCloseModals}
-            >
-              FAQ
-            </a>
-          </nav>
-        )}
-      </header>
+      <Navbar />
       <Background />
       {/* <a className="nav-return-top" href="#">
         Return to Top
